+++ src/NodeShared.cc
@@ -28,9 +28,11 @@
 #include <iostream>
 #include <map>
 #include <mutex>
+#include <shared_mutex>  //NOLINT
 #include <string>
 #include <thread>
 #include <vector>
+#include <unordered_map>
 
 // TODO(anyone): Remove after fixing the warnings.
 #ifdef _MSC_VER
@@ -91,7 +93,7 @@ bool userPass(std::string &_user, std::string &_pass)
 
 //////////////////////////////////////////////////
 // Helper to send messages
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
 int sendHelper(zmq::socket_t &_pub, const std::string &_data,
     const zmq::send_flags &_type)
 {
@@ -109,7 +111,7 @@ int sendHelper(zmq::socket_t &_pub, const std::string &_data, int _type)
 {
   zmq::message_t msg(_data.data(), _data.size());
 
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
   zmq::send_flags flags = zmq::send_flags::none;
   switch (_type)
   {
@@ -136,7 +138,7 @@ std::string receiveHelper(zmq::socket_t &_socket)
 {
   zmq::message_t msg(0);
 
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
   if (!_socket.recv(msg))
 #else
   if (!_socket.recv(&msg, 0))
@@ -152,7 +154,7 @@ std::string receiveHelper(zmq::socket_t &_socket)
 void sendAuthErrorHelper(zmq::socket_t &_socket, const std::string &_err)
 {
   std::cerr << _err << std::endl;
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
   sendHelper(_socket, "400", zmq::send_flags::sndmore);
   sendHelper(_socket, _err, zmq::send_flags::sndmore);
   sendHelper(_socket, "", zmq::send_flags::sndmore);
@@ -168,27 +170,45 @@ void sendAuthErrorHelper(zmq::socket_t &_socket, const std::string &_err)
 //////////////////////////////////////////////////
 NodeShared *NodeShared::Instance()
 {
-#ifdef _MSC_VER
-  // If we compile ign-transport as a shared library on Windows, we should
-  // never destruct NodeShared, unfortunately. It seems that WinSock does
-  // not behave well during the DLL teardown phase as a program exits, and
-  // this will confuse the ZeroMQ library into thinking that WinSock
-  // misbehaved, causing an assertion in ZeroMQ to fail and throw an exception
-  // while the program exits. This is a known issue:
-  //
-  // https://github.com/zeromq/libzmq/issues/1144
-  //
-  // An easy way of dodging this issue is to never destruct NodeShared. The
-  // Operating System will take care of cleaning up its resources when the
-  // application exits. We may want to consider a more elegant solution in
-  // the future. The zsys_shutdown() function in the czmq library may be able
-  // to provide some inspiration for solving this more cleanly.
-  static NodeShared *instance = new NodeShared();
-  return instance;
-#else
-  static NodeShared instance;
-  return &instance;
-#endif
+  // Create an instance of NodeShared per process so the ZMQ context
+  // is not shared between different processes.
+
+  static std::shared_mutex mutex;
+  static std::unordered_map<unsigned int, NodeShared*> nodeSharedMap;
+
+  // Get current process ID.
+  auto pid = getProcessId();
+
+  // Check if there's already a NodeShared instance for this process.
+  // Use a shared_lock so multiple threads can read simultaneously.
+  // This will only block if there's another thread locking exclusively
+  // for writing. Since most of the time threads will be reading,
+  // we make the read operation faster at the expense of making the write
+  // operation slower. Use exceptions for their zero-cost when successful.
+  try
+  {
+    std::shared_lock readLock(mutex);
+    return nodeSharedMap.at(pid);
+  }
+  catch (...)
+  {
+    // Multiple threads from the same process could have arrived here
+    // simultaneously, so after locking, we need to make sure that there's
+    // not an already constructed NodeShared instance for this process.
+    std::lock_guard writeLock(mutex);
+
+    auto iter = nodeSharedMap.find(pid);
+    if (iter != nodeSharedMap.end())
+    {
+      // There's already an instance for this process, return it.
+      return iter->second;
+    }
+
+    // No instance, construct a new one.
+    auto ret = nodeSharedMap.insert({pid, new NodeShared});
+    assert(ret.second);  // Insert operation should be successful.
+    return ret.first->second;
+  }
 }
 
 //////////////////////////////////////////////////
@@ -331,7 +351,7 @@ bool NodeShared::Publish(
     // Send the messages
     std::lock_guard<std::recursive_mutex> lock(this->mutex);
 
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
     this->dataPtr->publisher->send(msg0, zmq::send_flags::sndmore);
     this->dataPtr->publisher->send(msg1, zmq::send_flags::sndmore);
     this->dataPtr->publisher->send(msg2, zmq::send_flags::sndmore);
@@ -367,7 +387,7 @@ void NodeShared::RecvMsgUpdate()
 
     try
     {
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       if (!this->dataPtr->subscriber->recv(msg))
 #else
       if (!this->dataPtr->subscriber->recv(&msg, 0))
@@ -376,7 +396,7 @@ void NodeShared::RecvMsgUpdate()
       topic = std::string(reinterpret_cast<char *>(msg.data()), msg.size());
 
       // TODO(caguero): Use this as extra metadata for the subscriber.
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       if (!this->dataPtr->subscriber->recv(msg))
 #else
       if (!this->dataPtr->subscriber->recv(&msg, 0))
@@ -384,7 +404,7 @@ void NodeShared::RecvMsgUpdate()
         return;
       // sender = std::string(reinterpret_cast<char *>(msg.data()), msg.size());
 
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       if (!this->dataPtr->subscriber->recv(msg))
 #else
       if (!this->dataPtr->subscriber->recv(&msg, 0))
@@ -392,7 +412,7 @@ void NodeShared::RecvMsgUpdate()
         return;
       data = std::string(reinterpret_cast<char *>(msg.data()), msg.size());
 
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       if (!this->dataPtr->subscriber->recv(msg))
 #else
       if (!this->dataPtr->subscriber->recv(&msg, 0))
@@ -574,14 +594,14 @@ void NodeShared::RecvSrvRequest()
 
     try
     {
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       if (!this->dataPtr->replier->recv(msg))
 #else
       if (!this->dataPtr->replier->recv(&msg, 0))
 #endif
         return;
 
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       if (!this->dataPtr->replier->recv(msg))
 #else
       if (!this->dataPtr->replier->recv(&msg, 0))
@@ -589,7 +609,7 @@ void NodeShared::RecvSrvRequest()
         return;
       topic = std::string(reinterpret_cast<char *>(msg.data()), msg.size());
 
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       if (!this->dataPtr->replier->recv(msg))
 #else
       if (!this->dataPtr->replier->recv(&msg, 0))
@@ -597,7 +617,7 @@ void NodeShared::RecvSrvRequest()
         return;
       sender = std::string(reinterpret_cast<char *>(msg.data()), msg.size());
 
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       if (!this->dataPtr->replier->recv(msg))
 #else
       if (!this->dataPtr->replier->recv(&msg, 0))
@@ -605,7 +625,7 @@ void NodeShared::RecvSrvRequest()
         return;
       dstId = std::string(reinterpret_cast<char *>(msg.data()), msg.size());
 
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       if (!this->dataPtr->replier->recv(msg))
 #else
       if (!this->dataPtr->replier->recv(&msg, 0))
@@ -613,7 +633,7 @@ void NodeShared::RecvSrvRequest()
         return;
       nodeUuid = std::string(reinterpret_cast<char *>(msg.data()), msg.size());
 
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       if (!this->dataPtr->replier->recv(msg))
 #else
       if (!this->dataPtr->replier->recv(&msg, 0))
@@ -621,7 +641,7 @@ void NodeShared::RecvSrvRequest()
         return;
       reqUuid = std::string(reinterpret_cast<char *>(msg.data()), msg.size());
 
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       if (!this->dataPtr->replier->recv(msg))
 #else
       if (!this->dataPtr->replier->recv(&msg, 0))
@@ -629,7 +649,7 @@ void NodeShared::RecvSrvRequest()
         return;
       req = std::string(reinterpret_cast<char *>(msg.data()), msg.size());
 
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       if (!this->dataPtr->replier->recv(msg))
 #else
       if (!this->dataPtr->replier->recv(&msg, 0))
@@ -637,7 +657,7 @@ void NodeShared::RecvSrvRequest()
         return;
       reqType = std::string(reinterpret_cast<char *>(msg.data()), msg.size());
 
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       if (!this->dataPtr->replier->recv(msg))
 #else
       if (!this->dataPtr->replier->recv(&msg, 0))
@@ -700,7 +720,7 @@ void NodeShared::RecvSrvRequest()
 
       response.rebuild(dstId.size());
       memcpy(response.data(), dstId.data(), dstId.size());
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       this->dataPtr->replier->send(response, zmq::send_flags::sndmore);
 #else
       this->dataPtr->replier->send(response, ZMQ_SNDMORE);
@@ -708,7 +728,7 @@ void NodeShared::RecvSrvRequest()
 
       response.rebuild(topic.size());
       memcpy(response.data(), topic.data(), topic.size());
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       this->dataPtr->replier->send(response, zmq::send_flags::sndmore);
 #else
       this->dataPtr->replier->send(response, ZMQ_SNDMORE);
@@ -716,7 +736,7 @@ void NodeShared::RecvSrvRequest()
 
       response.rebuild(nodeUuid.size());
       memcpy(response.data(), nodeUuid.data(), nodeUuid.size());
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       this->dataPtr->replier->send(response, zmq::send_flags::sndmore);
 #else
       this->dataPtr->replier->send(response, ZMQ_SNDMORE);
@@ -724,7 +744,7 @@ void NodeShared::RecvSrvRequest()
 
       response.rebuild(reqUuid.size());
       memcpy(response.data(), reqUuid.data(), reqUuid.size());
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       this->dataPtr->replier->send(response, zmq::send_flags::sndmore);
 #else
       this->dataPtr->replier->send(response, ZMQ_SNDMORE);
@@ -732,7 +752,7 @@ void NodeShared::RecvSrvRequest()
 
       response.rebuild(rep.size());
       memcpy(response.data(), rep.data(), rep.size());
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       this->dataPtr->replier->send(response, zmq::send_flags::sndmore);
 #else
       this->dataPtr->replier->send(response, ZMQ_SNDMORE);
@@ -740,7 +760,7 @@ void NodeShared::RecvSrvRequest()
 
       response.rebuild(resultStr.size());
       memcpy(response.data(), resultStr.data(), resultStr.size());
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       this->dataPtr->replier->send(response, zmq::send_flags::none);
 #else
       this->dataPtr->replier->send(response, 0);
@@ -780,14 +800,14 @@ void NodeShared::RecvSrvResponse()
 
     try
     {
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       if (!this->dataPtr->responseReceiver->recv(msg))
 #else
       if (!this->dataPtr->responseReceiver->recv(&msg, 0))
 #endif
         return;
 
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       if (!this->dataPtr->responseReceiver->recv(msg))
 #else
       if (!this->dataPtr->responseReceiver->recv(&msg, 0))
@@ -795,7 +815,7 @@ void NodeShared::RecvSrvResponse()
         return;
       topic = std::string(reinterpret_cast<char *>(msg.data()), msg.size());
 
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       if (!this->dataPtr->responseReceiver->recv(msg))
 #else
       if (!this->dataPtr->responseReceiver->recv(&msg, 0))
@@ -803,7 +823,7 @@ void NodeShared::RecvSrvResponse()
         return;
       nodeUuid = std::string(reinterpret_cast<char *>(msg.data()), msg.size());
 
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       if (!this->dataPtr->responseReceiver->recv(msg))
 #else
       if (!this->dataPtr->responseReceiver->recv(&msg, 0))
@@ -811,7 +831,7 @@ void NodeShared::RecvSrvResponse()
         return;
       reqUuid = std::string(reinterpret_cast<char *>(msg.data()), msg.size());
 
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       if (!this->dataPtr->responseReceiver->recv(msg))
 #else
       if (!this->dataPtr->responseReceiver->recv(&msg, 0))
@@ -819,7 +839,7 @@ void NodeShared::RecvSrvResponse()
         return;
       rep = std::string(reinterpret_cast<char *>(msg.data()), msg.size());
 
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
       if (!this->dataPtr->responseReceiver->recv(msg))
 #else
       if (!this->dataPtr->responseReceiver->recv(&msg, 0))
@@ -953,7 +973,7 @@ void NodeShared::SendPendingRemoteReqs(const std::string &_topic,
 
         msg.rebuild(responserId.size());
         memcpy(msg.data(), responserId.data(), responserId.size());
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
         this->dataPtr->requester->send(msg, zmq::send_flags::sndmore);
 #else
         this->dataPtr->requester->send(msg, ZMQ_SNDMORE);
@@ -961,7 +981,7 @@ void NodeShared::SendPendingRemoteReqs(const std::string &_topic,
 
         msg.rebuild(_topic.size());
         memcpy(msg.data(), _topic.data(), _topic.size());
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
         this->dataPtr->requester->send(msg, zmq::send_flags::sndmore);
 #else
         this->dataPtr->requester->send(msg, ZMQ_SNDMORE);
@@ -970,7 +990,7 @@ void NodeShared::SendPendingRemoteReqs(const std::string &_topic,
         msg.rebuild(this->myRequesterAddress.size());
         memcpy(msg.data(), this->myRequesterAddress.data(),
           this->myRequesterAddress.size());
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
         this->dataPtr->requester->send(msg, zmq::send_flags::sndmore);
 #else
         this->dataPtr->requester->send(msg, ZMQ_SNDMORE);
@@ -979,7 +999,7 @@ void NodeShared::SendPendingRemoteReqs(const std::string &_topic,
         std::string myId = this->responseReceiverId.ToString();
         msg.rebuild(myId.size());
         memcpy(msg.data(), myId.data(), myId.size());
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
         this->dataPtr->requester->send(msg, zmq::send_flags::sndmore);
 #else
         this->dataPtr->requester->send(msg, ZMQ_SNDMORE);
@@ -987,7 +1007,7 @@ void NodeShared::SendPendingRemoteReqs(const std::string &_topic,
 
         msg.rebuild(nodeUuid.size());
         memcpy(msg.data(), nodeUuid.data(), nodeUuid.size());
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
         this->dataPtr->requester->send(msg, zmq::send_flags::sndmore);
 #else
         this->dataPtr->requester->send(msg, ZMQ_SNDMORE);
@@ -995,7 +1015,7 @@ void NodeShared::SendPendingRemoteReqs(const std::string &_topic,
 
         msg.rebuild(reqUuid.size());
         memcpy(msg.data(), reqUuid.data(), reqUuid.size());
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
         this->dataPtr->requester->send(msg, zmq::send_flags::sndmore);
 #else
         this->dataPtr->requester->send(msg, ZMQ_SNDMORE);
@@ -1003,7 +1023,7 @@ void NodeShared::SendPendingRemoteReqs(const std::string &_topic,
 
         msg.rebuild(data.size());
         memcpy(msg.data(), data.data(), data.size());
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
         this->dataPtr->requester->send(msg, zmq::send_flags::sndmore);
 #else
         this->dataPtr->requester->send(msg, ZMQ_SNDMORE);
@@ -1011,7 +1031,7 @@ void NodeShared::SendPendingRemoteReqs(const std::string &_topic,
 
         msg.rebuild(_reqType.size());
         memcpy(msg.data(), _reqType.data(), _reqType.size());
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
         this->dataPtr->requester->send(msg, zmq::send_flags::sndmore);
 #else
         this->dataPtr->requester->send(msg, ZMQ_SNDMORE);
@@ -1019,7 +1039,7 @@ void NodeShared::SendPendingRemoteReqs(const std::string &_topic,
 
         msg.rebuild(_repType.size());
         memcpy(msg.data(), _repType.data(), _repType.size());
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
         this->dataPtr->requester->send(msg, zmq::send_flags::none);
 #else
         this->dataPtr->requester->send(msg, 0);
@@ -1068,8 +1088,12 @@ void NodeShared::OnNewConnection(const MessagePublisher &_pub)
       this->dataPtr->subscriber->connect(addr.c_str());
 
     // Add a new filter for the topic.
+#ifdef IGN_CPPZMQ_POST_4_7_0
+    this->dataPtr->subscriber->set(zmq::sockopt::subscribe, topic);
+#else
     this->dataPtr->subscriber->setsockopt(ZMQ_SUBSCRIBE,
         topic.data(), topic.size());
+#endif
 
     // Register the new connection with the publisher.
     this->connections.AddPublisher(_pub);
@@ -1257,10 +1281,14 @@ bool NodeShared::InitializeSockets()
     // Initialize security
     this->dataPtr->SecurityInit();
 
-    char bindEndPoint[1024];
+
     int lingerVal = 0;
+#ifdef IGN_CPPZMQ_POST_4_7_0
+    this->dataPtr->publisher->set(zmq::sockopt::linger, lingerVal);
+#else
     this->dataPtr->publisher->setsockopt(ZMQ_LINGER,
         &lingerVal, sizeof(lingerVal));
+#endif
 
     // Set the capacity of the buffer for receiving messages.
     std::string ignRcvHwm;
@@ -1293,8 +1321,12 @@ bool NodeShared::InitializeSockets()
                   << std::endl;
       }
     }
+#ifdef IGN_CPPZMQ_POST_4_7_0
+    this->dataPtr->subscriber->set(zmq::sockopt::rcvhwm, rcvQueueVal);
+#else
     this->dataPtr->subscriber->setsockopt(ZMQ_RCVHWM,
           &rcvQueueVal, sizeof(rcvQueueVal));
+#endif
 
     // Set the capacity of the buffer for sending messages.
     std::string ignSndHwm;
@@ -1327,6 +1359,34 @@ bool NodeShared::InitializeSockets()
                   << std::endl;
       }
     }
+#ifdef IGN_CPPZMQ_POST_4_7_0
+    this->dataPtr->publisher->set(zmq::sockopt::sndhwm, sndQueueVal);
+
+    this->dataPtr->publisher->bind(anyTcpEp.c_str());
+    this->myAddress =
+        this->dataPtr->publisher->get(zmq::sockopt::last_endpoint);
+
+    // ResponseReceiver socket listening in a random port.
+    std::string id = this->responseReceiverId.ToString();
+    this->dataPtr->responseReceiver->set(zmq::sockopt::routing_id, id);
+    this->dataPtr->responseReceiver->bind(anyTcpEp.c_str());
+    this->myRequesterAddress = this->dataPtr->responseReceiver->get(
+        zmq::sockopt::last_endpoint);
+
+    // Replier socket listening in a random port.
+    id = this->replierId.ToString();
+    this->dataPtr->replier->set(zmq::sockopt::routing_id, id);
+    int routeOn = 1;
+    this->dataPtr->replier->set(zmq::sockopt::linger, lingerVal);
+    this->dataPtr->replier->set(zmq::sockopt::router_mandatory, routeOn);
+    this->dataPtr->replier->bind(anyTcpEp.c_str());
+    this->myReplierAddress =
+        this->dataPtr->replier->get(zmq::sockopt::last_endpoint);
+
+    this->dataPtr->requester->set(zmq::sockopt::linger, lingerVal);
+    this->dataPtr->requester->set(zmq::sockopt::router_mandatory, routeOn);
+#else
+    char bindEndPoint[1024];
     this->dataPtr->publisher->setsockopt(ZMQ_SNDHWM,
         &sndQueueVal, sizeof(sndQueueVal));
 
@@ -1361,6 +1421,7 @@ bool NodeShared::InitializeSockets()
         &lingerVal, sizeof(lingerVal));
     this->dataPtr->requester->setsockopt(ZMQ_ROUTER_MANDATORY, &RouteOn,
       sizeof(RouteOn));
+#endif
   }
   catch(const zmq::error_t& ze)
   {
@@ -1396,10 +1457,14 @@ bool NodeShared::AdvertisePublisher(const ServicePublisher &_publisher)
 int NodeShared::RcvHwm()
 {
   int rcvHwm;
-  size_t rcvHwmSize = sizeof(rcvHwm);
   try
   {
+#ifdef IGN_CPPZMQ_POST_4_7_0
+    rcvHwm = this->dataPtr->subscriber->get(zmq::sockopt::rcvhwm);
+#else
+    size_t rcvHwmSize = sizeof(rcvHwm);
     this->dataPtr->subscriber->getsockopt(ZMQ_RCVHWM, &rcvHwm, &rcvHwmSize);
+#endif
   }
   catch (zmq::error_t &_e)
   {
@@ -1413,10 +1478,14 @@ int NodeShared::RcvHwm()
 int NodeShared::SndHwm()
 {
   int sndHwm;
-  size_t sndHwmSize = sizeof(sndHwm);
   try
   {
+#ifdef IGN_CPPZMQ_POST_4_7_0
+    sndHwm = this->dataPtr->publisher->get(zmq::sockopt::sndhwm);
+#else
+    size_t sndHwmSize = sizeof(sndHwm);
     this->dataPtr->publisher->getsockopt(ZMQ_SNDHWM, &sndHwm, &sndHwmSize);
+#endif
   }
   catch (zmq::error_t &_e)
   {
@@ -1512,8 +1581,13 @@ void NodeSharedPrivate::SecurityOnNewConnection()
   // See issue #74
   if (userPass(user, pass))
   {
+#ifdef IGN_CPPZMQ_POST_4_7_0
+    this->subscriber->set(zmq::sockopt::plain_username, user);
+    this->subscriber->set(zmq::sockopt::plain_password, pass);
+#else
     this->subscriber->setsockopt(ZMQ_PLAIN_USERNAME, user.c_str(), user.size());
     this->subscriber->setsockopt(ZMQ_PLAIN_PASSWORD, pass.c_str(), pass.size());
+#endif
   }
 }
 
@@ -1531,11 +1605,16 @@ void NodeSharedPrivate::SecurityInit()
 
     int asPlainSecurityServer = static_cast<int>(
         ZmqPlainSecurityServerOptions::ZMQ_PLAIN_SECURITY_SERVER_ENABLED);
+
+#ifdef IGN_CPPZMQ_POST_4_7_0
+    this->publisher->set(zmq::sockopt::plain_server, asPlainSecurityServer);
+    this->publisher->set(zmq::sockopt::zap_domain, kIgnAuthDomain);
+#else
     this->publisher->setsockopt(ZMQ_PLAIN_SERVER,
         &asPlainSecurityServer, sizeof(asPlainSecurityServer));
-
     this->publisher->setsockopt(ZMQ_ZAP_DOMAIN, kIgnAuthDomain,
         std::strlen(kIgnAuthDomain));
+#endif
   }
 }
 
@@ -1645,7 +1724,7 @@ void NodeSharedPrivate::AccessControlHandler()
           continue;
         }
 
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
         sendHelper(*sock, version, zmq::send_flags::sndmore);
         sendHelper(*sock, sequence, zmq::send_flags::sndmore);
 #else
@@ -1656,7 +1735,7 @@ void NodeSharedPrivate::AccessControlHandler()
         // Check the username and password
         if (givenUsername == user && givenPassword == pass)
         {
-#ifdef IGN_ZMQ_POST_4_4_0
+#ifdef IGN_ZMQ_POST_4_3_1
           sendHelper(*sock, "200", zmq::send_flags::sndmore);
           sendHelper(*sock, "OK", zmq::send_flags::sndmore);
           sendHelper(*sock, "anonymous", zmq::send_flags::sndmore);